LayoutTests/http/tests/security/xslt-external-entity-expected.txt

ALERT: document() /etc/passwd: /etc/hosts: /etc/group:
